% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wellchart}
\alias{wellchart}
\title{Function for dataloading}
\usage{
wellchart(data, wellname, ...)
}
\arguments{
\item{data}{R dataset name}

\item{wellname}{Name of the well you would like to plot.}
}
\description{
Function for dataloading
}
\examples{
wellchart(monthly, "NSD10")
}
\keyword{chart}
\keyword{data}
\keyword{wells}

