% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wloaddata}
\alias{wloaddata}
\title{Function for dataloading}
\usage{
wloaddata(filepath, dbtable = "T_PROD_MON", backfilename = "monthly.csv",
  ...)
}
\arguments{
\item{filepath}{Location of source file}

\item{dbtable}{Name of database table you would like to import.
Default is "T_PROD_MON"}

\item{backfilename}{Name of csv file where the copy of imported
data will be stored. Default value "monthly.csv"}
}
\description{
Function for dataloading
}
\examples{
wloaddata("Z:/TVS/DB/GREZINIAI.accdb")
}
\keyword{data}
\keyword{import}
\keyword{wells}

