% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wnewvars}
\alias{wnewvars}
\title{Function for creating additional variables (summation, grouping
and cumulatives of imported data)}
\usage{
wnewvars(data)
}
\arguments{
\item{data}{Name of dataframe in R}
}
\description{
Function for creating additional variables (summation, grouping
and cumulatives of imported data)
}
\examples{
wnewvars(monthly)
}
\keyword{cumulatives}
\keyword{data}
\keyword{variables}
\keyword{wells}

