% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wsubbyfields}
\alias{wsubbyfields}
\title{Script which creates production data subsets for each oilfield}
\usage{
wsubbyfields()
}
\arguments{
\item{No}{input parameters for this script.}
}
\value{
output Created separate dataframes for particular fields. Dataframes
named as FIELDNAME.sub Data structure in dataframes is identical to monthly
dataframe structure.
}
\description{
Actually this is script not the function. It created for data splitting
automation.
When you ron this script, SQL sentence runs via sqldf() library and creates
separate dataframes for particular oilfieds.
All subsets are hardcoded. If you would like to add new field or remove
the existing one, you should edit this file.
Input dataframe also hardcoded as "monthly".
}
\examples{
wsubbyfields()
}
\keyword{keywords}
\keyword{no}

